+++ latex/UseProTeX.cmake
@@ -1,1 +1,2 @@
-set (protex_exe ${CMAKE_SOURCE_DIR}/config/protex)
+set (protex_exe ${CMAKE_CURRENT_LIST_DIR}/protex)
+message(STATUS "Found protex: ${protex_exe}")