+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Updated `UseLATEX` to version 2.7.2
+
 ### Fixed
 
 ### Removed
@@ -16,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added check to `esma.cmake` to ensure `CMAKE_INSTALL_PREFIX` is writable
+- Add `protex` script
 
 ## [3.18.0] - 2022-08-18